+++ MMOCoreORB/src/server/zone/objects/creature/buffs/PowerBoostBuffDurationEvent.h
@@ -19,7 +19,7 @@ namespace server {
 			ManagedWeakReference<Buff*> buffObject;
 
 		public:
-			PowerBoostBuffDurationEvent(CreatureObject* creature, Buff* buff) : Task((int) buff->getBuffDuration() * 1000) {
+			PowerBoostBuffDurationEvent(CreatureObject* creature, Buff* buff) : Task((int64) buff->getBuffDuration() * 1000) {
 				creatureObject = creature;
 				buffObject = buff;
 			}
